Added the interview video to the page and buy ticket buttons.  Moved the vide to be a component used by all pages.
